split out IO functions to utils/io.py
